BuildToolsSigning

+++ tools/CustomBuildTool/DynData.cs
@@ -259,12 +259,32 @@ typedef struct _KPH_DYNDATA
             string headerFile = $"{Build.BuildWorkingFolder}\\kphlib\\include\\kphdyn.h";
             string sourceFile = $"{Build.BuildWorkingFolder}\\kphlib\\kphdyn.c";
 
-            GenerateConfig(manifestFile, out byte[] config);
+            // Check for new or modified content. We don't want to touch the file if it's not needed.
+            {
+                string headerUpdateText = GenerateHeader();
+                string headerCurrentText = Utils.ReadAllText(headerFile);
 
-            Utils.WriteAllText(headerFile, GenerateHeader());
-            Program.PrintColorMessage($"Dynamic header -> {headerFile}", ConsoleColor.Cyan);
-            Utils.WriteAllText(sourceFile, GenerateSource(BytesToString(config)));
-            Program.PrintColorMessage($"Dynamic source -> {sourceFile}", ConsoleColor.Cyan);
+                if (!string.Equals(headerUpdateText, headerCurrentText, StringComparison.OrdinalIgnoreCase))
+                {
+                    Utils.WriteAllText(headerFile, headerUpdateText);
+                }
+
+                Program.PrintColorMessage($"Dynamic header -> {headerFile}", ConsoleColor.Cyan);
+            }
+
+            byte[] config = GenerateConfig(manifestFile);
+
+            {
+                var headerUpdateText = GenerateSource(BytesToString(config));
+                var headerCurrentText = Utils.ReadAllText(sourceFile);
+
+                if (!string.Equals(headerUpdateText, headerCurrentText, StringComparison.OrdinalIgnoreCase))
+                {
+                    Utils.WriteAllText(sourceFile, headerUpdateText);
+                }
+
+                Program.PrintColorMessage($"Dynamic source -> {sourceFile}", ConsoleColor.Cyan);
+            }
 
             if (string.IsNullOrWhiteSpace(OutDir))
                 return true;
@@ -272,11 +292,37 @@ typedef struct _KPH_DYNDATA
             string configFile = $"{OutDir}\\ksidyn.bin";
 
             if (File.Exists(configFile))
+            {
+                var configUpdateBytes = Utils.ReadAllBytes(configFile).AsSpan();
+                var configSigFileName = Path.ChangeExtension(configFile, ".sig");
+
+                if (configUpdateBytes.SequenceEqual(config) && File.Exists(configSigFileName))
+                {
+                    Program.PrintColorMessage($"Dynamic config -> {configFile}", ConsoleColor.Cyan);
+                    return true;
+                }
+
                 File.Delete(configFile);
-            Directory.CreateDirectory(OutDir);
-            Utils.WriteAllBytes(configFile, config);
-            Program.PrintColorMessage($"Dynamic config -> {configFile}", ConsoleColor.Cyan);
-            return Verify.CreateSigFile("kph", configFile, StrictChecks);
+                File.Delete(configSigFileName);
+                Utils.WriteAllBytes(configFile, config);
+
+                bool configFileSig = Verify.CreateSigFile("kph", configFile, StrictChecks);
+
+                Program.PrintColorMessage($"Dynamic config -> {configFile}", ConsoleColor.Cyan);
+
+                return configFileSig;
+            }
+            else
+            {
+                Directory.CreateDirectory(OutDir);
+                Utils.WriteAllBytes(configFile, config);
+
+                bool configFileSig = Verify.CreateSigFile("kph", configFile, StrictChecks);
+
+                Program.PrintColorMessage($"Dynamic config -> {configFile}", ConsoleColor.Cyan);
+
+                return configFileSig;
+            }
         }
 
         private static string GenerateHeader()
@@ -321,10 +367,7 @@ typedef struct _KPH_DYNDATA
             return sb.ToString();
         }
 
-        private static void GenerateConfig(
-            string ManifestFile,
-            out byte[] ConfigBytes
-            )
+        private static byte[] GenerateConfig(string ManifestFile)
         {
             var xml = new XmlDocument();
             var configs = new List<DynConfig>(10);
@@ -390,16 +433,16 @@ typedef struct _KPH_DYNDATA
 
                     var member = typeof(DynConfigArch).GetField(name);
 
-                    if (arch == null)
+                    if (string.IsNullOrEmpty(arch))
                     {
                         member.SetValueDirect(__makeref(config.ArchAMD64), Convert.ChangeType(value, member.FieldType));
                         member.SetValueDirect(__makeref(config.ArchARM64), Convert.ChangeType(value, member.FieldType));
                     }
-                    else if (arch == "amd64")
+                    else if (string.Equals(arch, "amd64", StringComparison.OrdinalIgnoreCase))
                     {
                         member.SetValueDirect(__makeref(config.ArchAMD64), Convert.ChangeType(value, member.FieldType));
                     }
-                    else if (arch == "arm64")
+                    else if (string.Equals(arch, "arm64", StringComparison.OrdinalIgnoreCase))
                     {
                         member.SetValueDirect(__makeref(config.ArchARM64), Convert.ChangeType(value, member.FieldType));
                     }
@@ -430,7 +473,7 @@ typedef struct _KPH_DYNDATA
                 writer.Write((uint)configs.Count);
                 writer.Write(MemoryMarshal.AsBytes(CollectionsMarshal.AsSpan(configs)));
 
-                ConfigBytes = stream.ToArray();
+                return stream.ToArray();
             }
         }
